Put KubernetesException somewhere better.
